fix(ci): stabilize wendaosearch live tests
