+++ packages/rust/crates/xiuxian-wendao-client/tests/unit/lint_run/syntax.rs
@@ -1,2 +1,2 @@
-#[path = "syntax_cases.rs"]
+#[path = "syntax_cases/mod.rs"]
 mod syntax_cases;
