+++ packages/rust/crates/xiuxian-wendao/tests/unit/search/queries/sql/information_schema.rs
@@ -47,11 +47,7 @@ async fn studio_sql_flight_provider_supports_information_schema_tables() {
     );
     assert_eq!(
         string_column_values(&response.batches[0], "table_type"),
-        vec![
-            "BASE TABLE".to_string(),
-            "BASE TABLE".to_string(),
-            "BASE TABLE".to_string(),
-        ]
+        vec!["VIEW".to_string(), "VIEW".to_string(), "VIEW".to_string()]
     );
 }
